[CDEC-416] Move `{Shutdown,StateLock,TimeWarp}` to `sinbin`
