+++ block/src/Pos/Block/BlockWorkMode.hs
@@ -26,11 +26,11 @@ import           Pos.Block.Types (LastKnownHeader, LastKnownHeaderTag,
 import           Pos.Core.Context (HasPrimaryKey)
 import           Pos.Infra.Communication.Protocol (Message)
 import           Pos.Infra.Recovery.Info (MonadRecoveryInfo)
-import           Pos.Infra.Shutdown.Class (HasShutdownContext)
-import           Pos.Infra.StateLock (StateLock, StateLockMetrics)
-import           Pos.Infra.Util.TimeWarp (CanJsonLog)
 import           Pos.Security.Params (SecurityParams)
+import           Pos.Sinbin.Shutdown.Class (HasShutdownContext)
+import           Pos.Sinbin.StateLock (StateLock, StateLockMetrics)
 import           Pos.Sinbin.Util.JsonLog.Events (MemPoolModifyReason)
+import           Pos.Sinbin.Util.TimeWarp (CanJsonLog)
 import           Pos.Txp (GenericTxpLocalData, MempoolExt, MonadTxpLocal,
                      TxpHolderTag)
 import           Pos.Update.Context (UpdateContext)
