+++ block/src/Pos/Block/Logic/Creation.hs
@@ -49,13 +49,13 @@ import           Pos.DB.Class (MonadDBRead)
 import           Pos.Delegation (DelegationVar, DlgPayload (..),
                      ProxySKBlockInfo, clearDlgMemPool, getDlgMempool)
 import           Pos.Exception (assertionFailed, reportFatalError)
-import           Pos.Infra.StateLock (Priority (..), StateLock,
-                     StateLockMetrics, modifyStateLock)
 import           Pos.Infra.Util.LogSafe (logInfoS)
 import           Pos.Lrc (HasLrcContext)
 import           Pos.Lrc.Context (lrcActionOnEpochReason)
 import qualified Pos.Lrc.DB as LrcDB
 import           Pos.Sinbin.Reporting (HasMisbehaviorMetrics, reportError)
+import           Pos.Sinbin.StateLock (Priority (..), StateLock,
+                     StateLockMetrics, modifyStateLock)
 import           Pos.Sinbin.Util.JsonLog.Events (MemPoolModifyReason (..))
 import           Pos.Ssc.Base (defaultSscPayload, stripSscPayload)
 import           Pos.Ssc.Logic (sscGetLocalPayload)
