+++ block/src/Pos/Block/Worker.hs
@@ -56,7 +56,6 @@ import           Pos.Infra.Recovery.Info (getSyncStatus, getSyncStatusK,
 import           Pos.Infra.Slotting (getSlotStartEmpatically, onNewSlot)
 import           Pos.Infra.Util.LogSafe (logDebugS, logInfoS, logWarningS)
 import           Pos.Infra.Util.TimeLimit (logWarningSWaitLinear)
-import           Pos.Infra.Util.TimeWarp (CanJsonLog (..))
 import qualified Pos.Lrc.DB as LrcDB (getLeadersForEpoch)
 import           Pos.Sinbin.Reporting (HasMisbehaviorMetrics,
                      MetricMonitor (..), MetricMonitorState, noReportMonitor,
@@ -66,6 +65,7 @@ import           Pos.Sinbin.Slotting (ActionTerminationPolicy (..),
                      OnNewSlotParams (..), currentTimeSlotting,
                      defaultOnNewSlotParams)
 import           Pos.Sinbin.Util.JsonLog.Events (jlCreatedBlock)
+import           Pos.Sinbin.Util.TimeWarp (CanJsonLog (..))
 
 import           Pos.Update.DB (getAdoptedBVData)
