+++ sinbin/cardano-sl-sinbin.cabal
@@ -24,11 +24,21 @@ library
                         Pos.Sinbin.Reporting.Methods
                         Pos.Sinbin.Reporting.Metrics
                         Pos.Sinbin.Reporting.Wlog
+
+                        Pos.Sinbin.Shutdown
+                        Pos.Sinbin.Shutdown.Class
+                        Pos.Sinbin.Shutdown.Logic
+                        Pos.Sinbin.Shutdown.Types
+
                         Pos.Sinbin.Slotting
                         Pos.Sinbin.Slotting.Class
                         Pos.Sinbin.Slotting.Types
                         Pos.Sinbin.Slotting.Util
                         Pos.Sinbin.Util.JsonLog.Events
+
+                        Pos.Sinbin.StateLock
+
+                        Pos.Sinbin.Util.TimeWarp
 
   other-modules:
                         Paths_cardano_sl_sinbin
