+++ sinbin/src/Pos/Sinbin/Shutdown.hs
@@ -1,0 +1,7 @@
+module Pos.Sinbin.Shutdown
+       ( module X
+       ) where
+
+import           Pos.Sinbin.Shutdown.Class as X
+import           Pos.Sinbin.Shutdown.Logic as X
+import           Pos.Sinbin.Shutdown.Types as X
