+++ ssc/src/Pos/Security/Params.hs
@@ -15,7 +15,7 @@ import           Serokell.Aeson.Options (defaultOptions)
 import qualified Text.Parsec as Parsec
 
 import           Pos.Core.Common (StakeholderId)
-import           Pos.Infra.Util.TimeWarp (NetworkAddress, addrParser)
+import           Pos.Sinbin.Util.TimeWarp (NetworkAddress, addrParser)
 import           Pos.Util.Util (aesonError, toAesonError)
 
 -- | Network attack settings (a part of the behavior config).
