+++ ssc/src/Pos/Security/Util.hs
@@ -7,9 +7,9 @@ import           Universum
 
 
 import           Pos.Core (StakeholderId)
-import           Pos.Infra.Util.TimeWarp (NetworkAddress)
 import           Pos.Security.Params (AttackTarget (..), AttackType (..),
                      SecurityParams (..))
+import           Pos.Sinbin.Util.TimeWarp (NetworkAddress)
 import           Pos.Util.Util (HasLens (..))
 
 shouldIgnoreAddress
