+++ ssc/src/Pos/Ssc/Mode.hs
@@ -13,12 +13,12 @@ import           System.Wlog (WithLogger)
 import           Pos.Core (HasPrimaryKey)
 import           Pos.DB.Class (MonadDB, MonadGState)
 import           Pos.Infra.Recovery.Info (MonadRecoveryInfo)
-import           Pos.Infra.Shutdown (HasShutdownContext)
-import           Pos.Infra.Util.TimeWarp (CanJsonLog)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Security.Params (SecurityParams)
 import           Pos.Sinbin.Reporting (MonadReporting)
+import           Pos.Sinbin.Shutdown (HasShutdownContext)
 import           Pos.Sinbin.Slotting (MonadSlots)
+import           Pos.Sinbin.Util.TimeWarp (CanJsonLog)
 import           Pos.Ssc.Configuration (HasSscConfiguration)
 import           Pos.Ssc.Mem (MonadSscMem)
 import           Pos.Ssc.Types (HasSscContext)
