+++ txp/src/Pos/Txp/Logic/Local.hs
@@ -35,10 +35,10 @@ import           Pos.Core.Txp (TxAux (..), TxId, TxUndo)
 import           Pos.Crypto (WithHash (..))
 import           Pos.DB.Class (MonadGState (..))
 import qualified Pos.DB.GState.Common as GS
-import           Pos.Infra.StateLock (Priority (..), StateLock,
-                     StateLockMetrics, withStateLock)
 import           Pos.Sinbin.Reporting (reportError)
 import           Pos.Sinbin.Slotting (MonadSlots (..))
+import           Pos.Sinbin.StateLock (Priority (..), StateLock,
+                     StateLockMetrics, withStateLock)
 import           Pos.Sinbin.Util.JsonLog.Events (MemPoolModifyReason (..))
 import           Pos.Txp.Configuration (tcAssetLockedSrcAddrs, txpConfiguration)
 import           Pos.Txp.Logic.Common (buildUtxo)
