+++ txp/src/Pos/Txp/MemState/Metrics.hs
@@ -12,7 +12,7 @@ import qualified System.Metrics as Metrics
 import qualified System.Metrics.Gauge as Metrics.Gauge
 import           System.Wlog (logDebug)
 
-import           Pos.Infra.StateLock (StateLockMetrics (..))
+import           Pos.Sinbin.StateLock (StateLockMetrics (..))
 import           Pos.Sinbin.Util.JsonLog.Events (JLEvent (..), JLMemPool (..),
                      MemPoolModifyReason (..))
 import           Pos.System.Metrics.Constants (withCardanoNamespace)
