+++ update/src/Pos/Update/Logic/Local.hs
@@ -40,9 +40,9 @@ import           Pos.Core.Update (UpId, UpdatePayload (..), UpdateProposal,
 import           Pos.Crypto (PublicKey, shortHashF)
 import           Pos.DB.Class (MonadDBRead)
 import qualified Pos.DB.GState.Common as DB
-import           Pos.Infra.StateLock (StateLock)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Sinbin.Reporting (MonadReporting)
+import           Pos.Sinbin.StateLock (StateLock)
 import           Pos.Update.Configuration (HasUpdateConfiguration)
 import           Pos.Update.Context (UpdateContext (..))
 import qualified Pos.Update.DB as DB
