+++ update/src/Pos/Update/MemState/Functions.hs
@@ -13,7 +13,7 @@ import           Pos.Binary.Class (biSize)
 import           Pos.Core.Configuration ()
 import           Pos.Core.Update (UpdatePayload (..), UpdateVote (..))
 import           Pos.Crypto (PublicKey, hash)
-import           Pos.Infra.StateLock (Priority (..), StateLock,
+import           Pos.Sinbin.StateLock (Priority (..), StateLock,
                      withStateLockNoMetrics)
 import           Pos.Update.MemState.Types (MemPool (..))
 import           Pos.Update.Poll (LocalVotes)
