+++ update/src/Pos/Update/Mode.hs
@@ -14,11 +14,11 @@ import           UnliftIO (MonadUnliftIO)
 import           Pos.Core.Configuration ()
 import           Pos.DB.Class (MonadDB, MonadGState)
 import           Pos.Infra.Recovery.Info (MonadRecoveryInfo)
-import           Pos.Infra.Shutdown.Class (HasShutdownContext)
-import           Pos.Infra.StateLock (StateLock)
 import           Pos.Lrc.Context (HasLrcContext)
 import           Pos.Sinbin.Reporting (MonadReporting)
+import           Pos.Sinbin.Shutdown.Class (HasShutdownContext)
 import           Pos.Sinbin.Slotting (MonadSlots)
+import           Pos.Sinbin.StateLock (StateLock)
 import           Pos.Update.Configuration (HasUpdateConfiguration)
 import           Pos.Update.Context (UpdateContext)
 import           Pos.Update.Params (UpdateParams)
