+++ update/src/Pos/Update/Worker.hs
@@ -16,8 +16,8 @@ import           Pos.Core (SoftwareVersion (..))
 import           Pos.Core.Update (UpdateProposal (..))
 import           Pos.Infra.Diffusion.Types (Diffusion)
 import           Pos.Infra.Recovery.Info (recoveryCommGuard)
-import           Pos.Infra.Shutdown (triggerShutdown)
 import           Pos.Infra.Slotting.Util (onNewSlot)
+import           Pos.Sinbin.Shutdown (triggerShutdown)
 import           Pos.Sinbin.Slotting.Util (ActionTerminationPolicy (..),
                      OnNewSlotParams (..), defaultOnNewSlotParams)
 import           Pos.Update.Configuration (curSoftwareVersion)
